refactor: simplify serve command syntax
