+++ .vscode/launch.json
@@ -8,8 +8,7 @@
             "mode": "auto",
             "program": "${workspaceFolder}",
             "args": [
-                "serve",
-                "authz-server"
+                "serve"
             ],
         }
     ]
