+++ charts/generic-auth-server/README.md
@@ -62,7 +62,7 @@ helm install generic-auth-server --namespace kyverno --create-namespace generic-
 | containers.server.livenessProbe | object | See [values.yaml](values.yaml) | Liveness probe. The block is directly forwarded into the deployment, so you can use whatever livenessProbe configuration you want. ref: https://kubernetes.io/docs/tasks/configure-pod-container/configure-liveness-readiness-probes/ |
 | containers.server.readinessProbe | object | See [values.yaml](values.yaml) | Readiness Probe. The block is directly forwarded into the deployment, so you can use whatever readinessProbe configuration you want. ref: https://kubernetes.io/docs/tasks/configure-pod-container/configure-liveness-readiness-probes/ |
 | containers.server.ports | list | `[{"containerPort":9080,"name":"health","protocol":"TCP"},{"containerPort":9081,"name":"auth","protocol":"TCP"}]` | Container ports. |
-| containers.server.args | list | `["serve","authz-server","--health-address=:9080","--auth-address=:9081"]` | Container args. |
+| containers.server.args | list | `["serve","--health-address=:9080","--auth-address=:9081"]` | Container args. |
 | service.port | int | `9081` | Service port. |
 | service.type | string | `"ClusterIP"` | Service type. |
 | service.nodePort | string | `nil` | Service node port. Only used if `type` is `NodePort`. |
